merge updates

+++ client/src/components/DashUsers.jsx
@@ -34,6 +34,7 @@ import "react-circular-progressbar/dist/styles.css";
 import Profile from "../assets/add-pic.png";
 import {
   HiOutlineExclamationCircle,
+  HiInformationCircle,
   HiPlusCircle,
   HiUserAdd,
 } from "react-icons/hi";
@@ -44,7 +45,9 @@ export default function DashUsers() {
   const [users, setUsers] = useState([]);
   const [showMore, setShowMore] = useState(true);
   const [showModal, setShowModal] = useState(false);
+  const [showModalDeletelock, setShowModalDeletelock] = useState(false);
   const [userIdToDelete, setUserIdToDelete] = useState("");
+  const [showAlert, setShowAlert] = useState(false);
 
   const [openModal, setOpenModal] = useState(false);
   const [openModalEdit, setOpenModalEdit] = useState(false);
@@ -61,6 +64,8 @@ export default function DashUsers() {
 
   const [createUserError, setCreateUserError] = useState(null);
   const [createLoding, setCreateLoding] = useState(null);
+
+  const [errorMessage, setErrorMessage] = useState(null);
 
   const filePickerRef = useRef();
 
@@ -238,16 +243,28 @@ export default function DashUsers() {
   };
 
   const handleDeleteUser = async () => {
+    if (currentUser.id === userIdToDelete) {
+      setShowAlert(true);
+      setShowModal(false);
+      return;
+    }
+
     try {
       const res = await fetch(`/api/user/deleteuser/${userIdToDelete}`, {
         method: "DELETE",
       });
       const data = await res.json();
+      if(res.status == 400) {
+        setShowModalDeletelock(true);
+        setErrorMessage(data.message);
+        setShowModal(false);
+      }
       if (res.ok) {
         setUsers((prev) => prev.filter((user) => user._id !== userIdToDelete));
         setShowModal(false);
         fetchUsers();
-      } else {
+      }
+      else {
         console.log(data.message);
       }
     } catch (error) {
@@ -715,6 +732,13 @@ export default function DashUsers() {
             </motion.div>
           </Modal>
 
+          {showAlert && (
+            <Alert className="mb-3" color="failure" icon={HiInformationCircle}>
+              <span className="font-medium">Info alert!</span> You can't delete
+              user account you are currently logged in.
+            </Alert>
+          )}
+
           {currentUser.role == "Admin" && users.length > 0 ? (
             <>
               <Table hoverable className="shadow-md w-full">
@@ -810,6 +834,34 @@ export default function DashUsers() {
                     </Button>
                     <Button color="gray" onClick={() => setShowModal(false)}>
                       No, cancel
+                    </Button>
+                  </div>
+                </div>
+              </Modal.Body>
+            </motion.div>
+          </Modal>
+          <Modal
+            show={showModalDeletelock}
+            onClose={() => setShowModalDeletelock(false)}
+            popup
+            size="md"
+          >
+            <motion.div
+              initial={{ opacity: 0, scale: 0.9 }}
+              animate={{ opacity: 1, scale: 1 }}
+              exit={{ opacity: 0, scale: 0.9 }}
+              transition={{ duration: 0.3 }}
+            >
+              <Modal.Header />
+              <Modal.Body>
+                <div className="text-center">
+                  <HiOutlineExclamationCircle className="h-14 w-14 text-yellow-400 dark:text-gray-200 mb-4 mx-auto" />
+                  <h3 className="mb-5 text-lg text-gray-500 dark:text-gray-400">
+                    {errorMessage}
+                  </h3>
+                  <div className="flex justify-center gap-4">
+                    <Button color="blue" onClick={() => setShowModalDeletelock(false)}>
+                      Okay
                     </Button>
                   </div>
                 </div>
